+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## GV 1.60+:
 
+### Release 1.8.2
+- Fixed an issue from the previous release where overrides would not be loaded due to case sensitivity.
+
+### Release 1.8.1
+- Files will no longer be cached if less than 10 GB of free space is available on the drive.
+- The cache folder is no longer hidden (that was a bad idea).
+- The persistent data path (used for storing the cache and mappings) can now be changed in the config.
+- Overrides are case-insensitive now.
+
+### Release 1.8.0
+- Fixed the new arena introduced in GV 1.62 being broken.
+
 ### Release 1.7.14
 - Fixed an IndexOutOfRangeException when exporting characters.